sql/pgwire: unify pgwire/datum date parsing code
Previously, pgwire had a separate timestamp parsing routine from the one
used by normal datum parsing. This produced inconsistencies where some
timestamps could be parsed in normal SQL execution but not pgwire
prepared statement execution.

Now, pgwire delegates to `ParseDTimestamp` and `ParseDTimestampTZ`
instead of its own parser.

Date parsing is now unified with timestamp and timestamptz parsing as
well, eliminating another potential source of parse inconsistencies.

+++ pkg/acceptance/java_test.go
@@ -81,6 +81,15 @@ public class main {
 		if (!tsStr.equals("2015-05-07 18:20:00.0")) {
 			throw new Exception("unexpected value for ts: "+tsStr);
 		}
+
+		stmt = conn.prepareStatement("INSERT INTO test.f VALUES (?, ?)");
+		stmt.setInt(1, 1);
+		stmt.setTimestamp(2, new Timestamp(System.currentTimeMillis()));
+		res = stmt.executeUpdate();
+		if (res != 1) {
+		    throw new Exception("unexpected: INSERT reports " + res + " rows changed, expecting 1");
+		}
+
 		stmt = conn.prepareStatement("DROP TABLE test.f");
 		res = stmt.executeUpdate();
 		if (res != 0) {
